+++ packages/harmony_log/guide/simple.dart
@@ -3,7 +3,7 @@ import 'package:harmony_log/harmony_log.dart';
 void main() {
   final log = Log(
     id: LogId(),
-    output: LogOutput.redirectOnDebug(
+    child: LogOutput.redirectOnDebug(
       child: LogOutput.plain(
         format: LogPlainFormat.simple(),
         child: LogPlainOutput.console(),