+++ packages/harmony_log/CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 1.0.0
 
 - Initial release
+
+## 1.1.0
+
+- **BREAKING**: Using `child` in `Log` instead of `output`.
+- Now `Log` also implements `LogOutput`.